cleanup-account-list-item
