+++ ui/app/components/app/account-list-item/index.js
@@ -1,0 +1,1 @@
+export { default } from './account-list-item'
