+++ ui/app/components/app/index.scss
@@ -1,3 +1,4 @@
+@import 'account-list-item/index';
 @import 'account-menu/index';
 @import 'add-token-button/index';
 @import 'alerts/alerts';
