+++ ui/app/css/base-styles.scss
@@ -24,8 +24,6 @@ html {
   input:focus,
   textarea:focus,
   .unit-input__input,
-  .account-list-item__account-primary-balance,
-  .account-list-item__input,
   .currency-display__input {
     outline: none;
   }
