--- ui/app/css/itcss/components/account-dropdown.scss
@@ -1,104 +1,0 @@
-.account-dropdown-balance {
-  color: $dusty-gray;
-  line-height: 19px;
-}
-
-.account-dropdown-edit-button {
-  color: $dusty-gray;
-
-  &:hover {
-    color: $white;
-  }
-}
-
-.account-list-item {
-  &__top-row {
-    display: flex;
-    margin-top: 10px;
-    margin-left: 8px;
-    position: relative;
-  }
-
-  &__tooltip-wrapper {
-    left: -10px;
-  }
-
-  &__account-balances {
-    height: auto;
-    border: none;
-    background-color: transparent;
-    color: #9b9b9b;
-    margin-left: 34px;
-    margin-top: 4px;
-    position: relative;
-  }
-
-  &__primary-cached-container {
-    display: flex;
-  }
-
-  &__cached-star {
-    margin-left: 4px;
-  }
-
-  &__cached-balances {
-    div:first-of-type {
-      color: $web-orange;
-    }
-
-    div:last-of-type {
-      color: rgba(220, 153, 18, 0.6901960784313725);
-    }
-  }
-
-  &__account-name {
-    font-size: 16px;
-    margin-left: 8px;
-  }
-
-  &__icon {
-    position: absolute;
-    right: 12px;
-    top: 1px;
-  }
-
-  &__account-primary-balance,
-  &__account-secondary-balance {
-    line-height: 16px;
-    font-size: 12px;
-  }
-
-  &__balance-flag {
-    position: absolute;
-    top: 3px;
-    left: -8px;
-    color: $primary-blue;
-  }
-
-  &__account-primary-balance {
-    color: $scorpion;
-    border: none;
-  }
-
-  &__account-secondary-balance {
-    color: $dusty-gray;
-  }
-
-  &__account-address {
-    margin-left: 35px;
-    width: 80%;
-    overflow: hidden;
-    text-overflow: ellipsis;
-  }
-
-  &__dropdown {
-    &:hover {
-      background: rgba($alto, 0.2);
-      cursor: pointer;
-
-      input {
-        background: rgba($alto, 0.1);
-      }
-    }
-  }
-}
