+++ ui/app/css/itcss/components/index.scss
@@ -9,7 +9,6 @@
 @import './modal';
 @import './alert';
 @import './newui-sections';
-@import './account-dropdown';
 @import './send';
 @import './confirm';
 @import './loading-overlay';
