--- ui/app/pages/send/account-list-item/account-list-item.container.js
@@ -1,21 +1,0 @@
-import { connect } from 'react-redux'
-import {
-  getNativeCurrency,
-  getIsMainnet,
-  isBalanceCached,
-  getPreferences,
-} from '../../../selectors'
-import AccountListItem from './account-list-item.component'
-
-export default connect(mapStateToProps)(AccountListItem)
-
-function mapStateToProps (state) {
-  const { showFiatInTestnets } = getPreferences(state)
-  const isMainnet = getIsMainnet(state)
-
-  return {
-    nativeCurrency: getNativeCurrency(state),
-    balanceIsCached: isBalanceCached(state),
-    showFiat: (isMainnet || !!showFiatInTestnets),
-  }
-}
