--- ui/app/pages/send/account-list-item/index.js
@@ -1,1 +1,0 @@
-export { default } from './account-list-item.container'
